Moved package management to UV.
