+++ pyproject.toml
@@ -1,35 +1,22 @@
-[build-system]
-requires = ["setuptools>=42", "wheel"]
-build-backend = "setuptools.build_meta"
-
 [project]
-name = "cwrappy"
+name = "cwrap-py"
 version = "0.1.0"
-description = "A code comment header tool that uses figlet, written in python."
-authors = [
-    {name = "Tyler Durham", email = "tyler.durham.tn@gmail.com"}
-]
-license = {text = "MIT"}
+description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.6"
-dependencies = [
-  "click>=8.1.8",
-  "markdown-it-py>=3.0.0",
-  "mdurl>=0.1.2",
-  "pyperclip>=1.9.0",
-  "pyfiglet>=1.0.2",
-  "Pygments>=2.19.1",
-  "rich>=14.0.0",
-  "shellingham>=1.5.4",
-  "typer>=0.15.3",
-  "typing_extensions>=4.13.2"
+authors = [
+    { name = "TylerDurham", email = "TylerDurham@users.noreply.github.com" }
 ]
-
-[project.urls]
-"Homepage" = "https://github.com/TylerDurham/cwrap-python"
+requires-python = ">=3.13"
+dependencies = [
+    "pyfiglet>=1.0.2",
+    "pyperclip>=1.9.0",
+    "rich>=14.0.0",
+    "typer>=0.15.4",
+]
 
 [project.scripts]
-cwrappy = "cwrappy.cli:main"
+cwrap-py = "cwrap_py:main"
 
-[tool.setuptools.package-data]
-cwrappy = ["cwrappy/*.py"]
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
