+++ src/cwrap_py/__init__.py
@@ -1,0 +1,7 @@
+__version__ = "0.1.1"
+
+from cwrap_py import cli
+
+
+def main():
+    cli.run()
