+++ src/cwrap_pyold/__init__.py
@@ -1,0 +1,2 @@
+def main() -> None:
+    print("Hello from cwrap-py!")
